Move stores to same folder
