+++ app/dev_mode/views.py
@@ -2,8 +2,8 @@ import logging
 import os
 import time
 
+from app.data_model.metadata_store import MetaDataConstants
 from app.dev_mode.jwt_encoder import Encoder
-from app.metadata.metadata_store import MetaDataConstants
 from app.schema_loader.schema_loader import available_schemas
 
 from flask import abort
