+++ tests/app/templating/test_metadata_template_preprocessor.py
@@ -1,6 +1,6 @@
 from app.templating.metadata_template_preprocessor import MetaDataTemplatePreprocessor
 from app.authentication.user import User
-from app.metadata.metadata_store import MetaDataStore, MetaDataConstants
+from app.data_model.metadata_store import MetaDataStore, MetaDataConstants
 
 from tests.app.framework.sr_unittest import SurveyRunnerTestCase
